% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_weekday}
\alias{add_weekday}
\title{Add weekday labels}
\usage{
add_weekday(data, date_col = "date")
}
\arguments{
\item{data}{Data frame with a date column.}

\item{date_col}{Date column name.}
}
\value{
\code{data} with a \code{weekday} column.
}
\description{
Add weekday labels
}
