% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{filter_valid_appusage}
\alias{filter_valid_appusage}
\title{Filter APP Usage records to participants passing QC}
\usage{
filter_valid_appusage(data, qc_table)
}
\arguments{
\item{data}{Day-level APP Usage data.}

\item{qc_table}{Output from \code{qc_appusage_day()}.}
}
\value{
Filtered \code{data}.
}
\description{
Filter APP Usage records to participants passing QC
}
