% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{parse_appusage_filename}
\alias{parse_appusage_filename}
\title{Parse BIDS-like APP Usage output filenames}
\usage{
parse_appusage_filename(filename)
}
\arguments{
\item{filename}{File name or path.}
}
\value{
A list of key-value entities plus extension.
}
\description{
Parse BIDS-like APP Usage output filenames
}
