% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_ms_value}
\alias{parse_ms_value}
\title{Parse millisecond values}
\usage{
parse_ms_value(x)
}
\arguments{
\item{x}{Values to parse.}
}
\value{
Numeric vector.
}
\description{
Parses APP Usage millisecond fields, including \code{"T:"} prefixes and
scientific notation such as \code{1.1447863E7}.
}
