% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_as_datetime}
\alias{safe_as_datetime}
\title{Safely parse APP Usage datetimes}
\usage{
safe_as_datetime(x, tz = "Asia/Shanghai")
}
\arguments{
\item{x}{Datetime values such as \code{"2023-11-12 08:43:46:870"}.}

\item{tz}{Time zone.}
}
\value{
POSIXct vector.
}
\description{
Safely parse APP Usage datetimes
}
