% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_package_name}
\alias{standardize_package_name}
\title{Standardize Android package names}
\usage{
standardize_package_name(x)
}
\arguments{
\item{x}{Package names.}
}
\value{
Lowercase trimmed package names. The artificial \code{ALL} package is
standardized to \code{"ALL"}.
}
\description{
Standardize Android package names
}
