% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{build_appusage_filename}
\alias{build_appusage_filename}
\title{Build BIDS-like APP Usage output filenames}
\usage{
build_appusage_filename(
  participant_id,
  export_type,
  proc = 1,
  extension = "rda"
)
}
\arguments{
\item{participant_id}{Participant identifier.}

\item{export_type}{APP Usage export type.}

\item{proc}{Processing step, using \code{1}, \code{2}, or \code{3}.}

\item{extension}{File extension without leading dot.}
}
\value{
A file name.
}
\description{
Uses key-value entities joined by underscores, with each key and value joined
by a hyphen.
}
